nouveau: rename aux.c to auxiliary.c for reviewing it on Windows
When I tried to review the linux kernel on Windows in my laptop
and incidentally found that it failed to open the aux.c.

And Microsoft tells me:
(http://msdn.microsoft.com/en-us/library/aa365247.aspx)

> Do not use the following reserved names for the name of a file:
> CON, PRN, AUX, NUL, <snip...>, and LPT9. Also avoid these names
> followed immediately by an extension; for example, NUL.txt
> is not recommended.

The name "aux" is listed above. And it sometimes makes sense to
review linux on windows, so we rename the aux.c to auxiliary.c.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>